requirements.txt
# Bos Upety Bot PRO 24/7 - Dependencies
# Core Flask web framework
flask==2.3.3

# HTTP requests
requests==2.31.0

# Google AI (Gemini)
google-generativeai==0.3.2

# Google APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Google Sheets
gspread==5.12.0
oauth2client==4.1.3

# Data processing
pandas==2.1.3

# Task scheduling
schedule==1.2.0

# Environment variables
python-dotenv==1.0.0

# Excel support (optional)
openpyxl==3.1.2

# Additional utilities
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
